ml/requirements.txt
rembg>=2.0.69
flask>=3.0.0
pillow>=10.0.0
onnxruntime>=1.18.0